Quick and dirty conversion for export command
Before this commit, when wav files were given to the app and then
an export command was requested, the files were simply copied. However, the
app doesn't check at any point that the wav files are readable by the MPC
(meaning 44100Hz 16bit mono or stereo signed PCM in riff files). After this
commit, the export command converts the wav files that couldn't be read by
the MPC into files that can, eliminating in most case the "WRONG FILE" issue